refactor(web): extract shared OutcomeCard component and fix review findings
- Extract OutcomeCard, parseOutcomeFields, and regexes into OutcomeCard.tsx
- Replace VERDICT_COLORS dead code with KNOWN_VERDICTS Set
- Fix end-marker regex to also accept standalone --- (matches server behaviour)
- Fix magic number font-size: 10px → var(--text-xs) in .outcomeRawToggle
- Remove duplicated outcome CSS from RenderedContent and MarkdownContent modules
- Move detailed OutcomeCard tests to OutcomeCard.test.tsx; keep 2 integration
  tests per component file
