+++ web/src/modules/shared/components/SessionChat/OutcomeCard.module.css
@@ -1,0 +1,120 @@
+/* OutcomeCard — shared outcome block renderer styles */
+
+.outcomeCard {
+  margin-top: var(--space-3);
+  margin-bottom: var(--space-3);
+  border-radius: var(--radius-lg);
+  border: 1px solid var(--color-border-subtle);
+  background-color: color-mix(in srgb, var(--color-bg-secondary) 60%, transparent);
+  overflow: hidden;
+}
+
+.outcomeHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: var(--space-2) var(--space-4);
+  border-bottom: 1px solid var(--color-border-subtle);
+}
+
+.outcomeLabel {
+  font-size: var(--text-xs);
+  font-family: var(--font-mono);
+  color: var(--color-text-faint);
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+}
+
+.outcomeHeaderRight {
+  display: flex;
+  align-items: center;
+  gap: var(--space-2);
+}
+
+.outcomeBadge {
+  font-size: var(--text-xs);
+  font-family: var(--font-mono);
+  font-weight: 600;
+  padding: 1px 8px;
+  border-radius: var(--radius-full);
+  border: 1px solid;
+  text-transform: uppercase;
+  letter-spacing: 0.03em;
+}
+
+.outcomeBadge[data-verdict='approve'],
+.outcomeBadge[data-verdict='pass'] {
+  color: var(--color-accent-emerald);
+  border-color: var(--color-accent-emerald);
+}
+
+.outcomeBadge[data-verdict='retry'] {
+  color: var(--color-accent-amber);
+  border-color: var(--color-accent-amber);
+}
+
+.outcomeBadge[data-verdict='escalate'],
+.outcomeBadge[data-verdict='fail'] {
+  color: var(--color-accent-red);
+  border-color: var(--color-accent-red);
+}
+
+.outcomeBadge[data-verdict='unknown'] {
+  color: var(--color-text-secondary);
+  border-color: var(--color-text-secondary);
+}
+
+.outcomeRawToggle {
+  font-size: var(--text-xs);
+  font-family: var(--font-mono);
+  color: var(--color-text-faint);
+  background: none;
+  border: none;
+  cursor: pointer;
+  padding: 0;
+  transition: color var(--transition-fast);
+}
+
+.outcomeRawToggle:hover {
+  color: var(--color-text-muted);
+}
+
+.outcomeFields {
+  padding: var(--space-3) var(--space-4);
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-2);
+}
+
+.outcomeField {
+  display: flex;
+  gap: var(--space-3);
+  font-size: var(--text-xs);
+  line-height: 1.5;
+}
+
+.outcomeKey {
+  font-family: var(--font-mono);
+  color: var(--color-text-faint);
+  min-width: 10rem;
+  flex-shrink: 0;
+}
+
+.outcomeValue {
+  color: var(--color-text-secondary);
+  word-break: break-word;
+}
+
+.outcomeRaw {
+  border-top: 1px solid var(--color-border-subtle);
+  padding: var(--space-3) var(--space-4);
+}
+
+.outcomeRawYaml {
+  font-size: var(--text-xs);
+  font-family: var(--font-mono);
+  color: var(--color-text-muted);
+  white-space: pre-wrap;
+  word-break: break-word;
+  margin: 0;
+}
