+++ web/src/modules/shared/components/SessionChat/RenderedContent.test.tsx
@@ -428,113 +428,22 @@ describe('RenderedContent — complex scenarios', () => {
 });
 
 /* ================================================================== */
-/*  RenderedContent — Outcome blocks                                   */
+/*  RenderedContent — Outcome blocks (integration)                     */
 /* ================================================================== */
 
 describe('RenderedContent — outcome blocks', () => {
   const outcomeBlock = `---outcome---
 verdict: approve
-tests_passing: true
-scope_adherence: 0.95
-summary: Implementation complete with full test coverage
+summary: Implementation complete
 ---end---`;
 
   it('renders an outcome block as a card instead of raw text', () => {
     const { container } = render(<RenderedContent content={outcomeBlock} />);
 
-    // Should not contain the raw markers
     expect(screen.queryByText(/---outcome---/)).toBeNull();
     expect(screen.queryByText(/---end---/)).toBeNull();
-
-    // Should render as a card with the "Outcome" label
     expect(screen.getByText('Outcome')).toBeInTheDocument();
     expect(container.querySelector('[class*="outcomeCard"]')).toBeInTheDocument();
-  });
-
-  it('renders the verdict as a badge', () => {
-    render(<RenderedContent content={outcomeBlock} />);
-
-    expect(screen.getByText('approve')).toBeInTheDocument();
-  });
-
-  it('sets data-verdict="approve" on the badge for green coloring', () => {
-    const { container } = render(<RenderedContent content={outcomeBlock} />);
-
-    const badge = container.querySelector('[class*="outcomeBadge"]');
-    expect(badge).not.toBeNull();
-    expect(badge).toHaveAttribute('data-verdict', 'approve');
-  });
-
-  it('sets data-verdict="retry" for retry verdict', () => {
-    const content = '---outcome---\nverdict: retry\nsummary: needs more work\n---end---';
-    const { container } = render(<RenderedContent content={content} />);
-
-    const badge = container.querySelector('[class*="outcomeBadge"]');
-    expect(badge).toHaveAttribute('data-verdict', 'retry');
-  });
-
-  it('sets data-verdict="escalate" for escalate verdict', () => {
-    const content = '---outcome---\nverdict: escalate\nsummary: requires escalation\n---end---';
-    const { container } = render(<RenderedContent content={content} />);
-
-    const badge = container.querySelector('[class*="outcomeBadge"]');
-    expect(badge).toHaveAttribute('data-verdict', 'escalate');
-  });
-
-  it('sets data-verdict="unknown" for unrecognized verdicts', () => {
-    const content = '---outcome---\nverdict: custom_status\nsummary: unusual\n---end---';
-    const { container } = render(<RenderedContent content={content} />);
-
-    const badge = container.querySelector('[class*="outcomeBadge"]');
-    expect(badge).toHaveAttribute('data-verdict', 'unknown');
-  });
-
-  it('renders non-verdict fields as key-value pairs', () => {
-    render(<RenderedContent content={outcomeBlock} />);
-
-    expect(screen.getByText('tests_passing')).toBeInTheDocument();
-    expect(screen.getByText('true')).toBeInTheDocument();
-    expect(screen.getByText('scope_adherence')).toBeInTheDocument();
-    expect(screen.getByText('0.95')).toBeInTheDocument();
-    expect(screen.getByText('summary')).toBeInTheDocument();
-    expect(screen.getByText('Implementation complete with full test coverage')).toBeInTheDocument();
-  });
-
-  it('does not render verdict as a field row (only as badge)', () => {
-    render(<RenderedContent content={outcomeBlock} />);
-
-    // 'approve' should only appear once (as the badge), not as a field value
-    const approveElements = screen.getAllByText('approve');
-    expect(approveElements).toHaveLength(1);
-  });
-
-  it('shows "Show raw" button initially', () => {
-    render(<RenderedContent content={outcomeBlock} />);
-
-    expect(screen.getByText('Show raw')).toBeInTheDocument();
-    expect(screen.queryByText('Hide raw')).toBeNull();
-  });
-
-  it('shows raw YAML when "Show raw" is clicked', () => {
-    render(<RenderedContent content={outcomeBlock} />);
-
-    fireEvent.click(screen.getByText('Show raw'));
-
-    expect(screen.getByText('Hide raw')).toBeInTheDocument();
-    // Raw yaml content should now be visible
-    const pre = screen.getByText(/verdict: approve/);
-    expect(pre.tagName).toBe('PRE');
-  });
-
-  it('hides raw YAML when "Hide raw" is clicked', () => {
-    render(<RenderedContent content={outcomeBlock} />);
-
-    fireEvent.click(screen.getByText('Show raw'));
-    expect(screen.getByText('Hide raw')).toBeInTheDocument();
-
-    fireEvent.click(screen.getByText('Hide raw'));
-    expect(screen.getByText('Show raw')).toBeInTheDocument();
-    expect(screen.queryByText(/verdict: approve/)).toBeNull();
   });
 
   it('strips outcome markers from surrounding message text', () => {
@@ -544,24 +453,5 @@ summary: Implementation complete with full test coverage
     expect(screen.getByText('Before the outcome.')).toBeInTheDocument();
     expect(screen.getByText('After the outcome.')).toBeInTheDocument();
     expect(screen.getByText('Outcome')).toBeInTheDocument();
-  });
-
-  it('handles outcome block with no verdict gracefully', () => {
-    const content = '---outcome---\nsummary: analysis complete\n---end---';
-    const { container } = render(<RenderedContent content={content} />);
-
-    expect(screen.getByText('Outcome')).toBeInTheDocument();
-    // No badge when verdict is absent
-    expect(container.querySelector('[class*="outcomeBadge"]')).toBeNull();
-    expect(screen.getByText('analysis complete')).toBeInTheDocument();
-  });
-
-  it('renders multiple outcome blocks', () => {
-    const block2 = '---outcome---\nverdict: fail\nsummary: second check\n---end---';
-    const content = `${outcomeBlock}\n\n${block2}`;
-    const { container } = render(<RenderedContent content={content} />);
-
-    const cards = container.querySelectorAll('[class*="outcomeCard"]');
-    expect(cards).toHaveLength(2);
   });
 });
